Convert the frame from uint8 to float
Convert the frame from uint8 to float in the range 0-255 in the beginning,
so all filters can assume that they are able to multiply, add substract
without numeric overflow or underflow.

This also improves the performance, as the frame needs to be converted
less often between uint8 and other data types.

+++ filters/stripes.py
@@ -6,7 +6,7 @@ roll_y = 0
 def stripes(width=5, intensity=10.0, speed=1, *args, **kwargs):
     global roll_y
     roll_y = (roll_y + speed) % (width*2)
-    frame = kwargs['frame'].astype(np.float)
+    frame = kwargs['frame']
     for i in range(width):
         frame[i+roll_y+0::2*width,:,:3] -= intensity
         frame[i+roll_y+width::2*width,:,:3] += intensity
